+++ services/loader.sh
@@ -4,3 +4,4 @@
 #
 
 source $DIR/services/nginx-proxy.sh
+source $DIR/services/adminer.sh